LayoutTests/fast/frames/sandboxed-iframe-attribute-parsing-expected.txt

CONSOLE MESSAGE: line 113: Error while parsing the 'sandbox' attribute: 'För', 'var', 'vers,', 'jag', 'gör,', 'Lovar', 'du', 'en', 'kyss', 'mig', 'giva;', 'Arket', 'fullt', 'jag', 'borde', 'skriva,', 'Mindre', 'har', 'jag', 'skrivit', 'för.', 'Men', 'man', 'måste', 'hålla', 'måtta,', 'Jag', 'med', 'vers,', 'med', 'kyssar', 'du.', 'Låt', 'mig', 'räkna:', 'Där', 'är', 'sju!', 'En', 'därtill', 'det', 'gör', 'mig', 'åtta.', 'Numro', 'åtta', 'är', 'fatal,', 'Greklands', 'sångmör', 'voro', 'nio,', 'Och', 'en', 'svensk', 'därtill', 'gör', 'tio.', '—', 'Elva', 'var', 'apostelns', 'tal,', 'Ty', 'jag', 'räknar', 'icke', 'Judas,', 'Honom,', 'som', 'i', 'vänners', 'lag', 'Kysste', 'falskt;', 'det', 'gör', 'ej', 'jag,', 'Helst', 'när', 'vackra', 'läppar', 'bjudas.', 'Huru', 'står', 'min', 'räkning', 'här?', 'Aderton;', 'det', 'är', 'dock', 'något.', 'Nitton', '—', 'rimmet', 'gör', 'besvär,', 'Därföre', 'jag', 'fyller', 'tjoget.', 'Strofen', 'är', 'ej', 'full', 'som', 'jag,', 'In', 'i', 'hamnen', 'vill', 'jag', 'styra,', 'Därföre', 'till', 'godo', 'tag', 'Denna', 'gång', 'med', 'tjugofyra.', ''Kyssarna'', '('The', 'kisses'),', 'Esaias', 'Tegnér,', '1782-1846', 'int', 'main(void)', '{', 'return', '0;', '}' are invalid sandbox flags.
CONSOLE MESSAGE: line 131: Error while parsing the 'sandbox' attribute: 'allowscripts' is an invalid sandbox flag.
CONSOLE MESSAGE: line 135: Error while parsing the 'sandbox' attribute: 'allows-cripts' is an invalid sandbox flag.
CONSOLE MESSAGE: line 139: Error while parsing the 'sandbox' attribute: '-allow-scripts' is an invalid sandbox flag.
CONSOLE MESSAGE: line 143: Error while parsing the 'sandbox' attribute: 'allow_scripts' is an invalid sandbox flag.
CONSOLE MESSAGE: line 147: Error while parsing the 'sandbox' attribute: 'allowScripts' is an invalid sandbox flag.
CONSOLE MESSAGE: line 151: Error while parsing the 'sandbox' attribute: 'aallow-scripts' is an invalid sandbox flag.
CONSOLE MESSAGE: line 155: Error while parsing the 'sandbox' attribute: 'allow-scriptss' is an invalid sandbox flag.
CONSOLE MESSAGE: Blocked form submission to 'javascript:top.disallowedFormSubmitted();' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
CONSOLE MESSAGE: Blocked form submission to 'javascript:top.disallowedFormSubmitted();' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
CONSOLE MESSAGE: Blocked form submission to 'javascript:top.disallowedFormSubmitted();' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
CONSOLE MESSAGE: Blocked form submission to 'javascript:top.disallowedFormSubmitted();' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
CONSOLE MESSAGE: Blocked form submission to 'javascript:top.disallowedFormSubmitted();' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
CONSOLE MESSAGE: Blocked script execution in 'sandboxed-iframe-attribute-parsing-disallowed.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: Blocked script execution in 'sandboxed-iframe-attribute-parsing-disallowed.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: Blocked script execution in 'sandboxed-iframe-attribute-parsing-disallowed.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: Blocked script execution in 'sandboxed-iframe-attribute-parsing-disallowed.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: Blocked script execution in 'sandboxed-iframe-attribute-parsing-disallowed.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: Blocked script execution in 'sandboxed-iframe-attribute-parsing-disallowed.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: Blocked script execution in 'sandboxed-iframe-attribute-parsing-disallowed.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: Blocked script execution in 'sandboxed-iframe-attribute-parsing-disallowed.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
This test case verifies the parsing of the iframe sandbox attribute. Two sets of iframes are used: one where scripting is allowed, and another one where it is disallowed. The test verifies that the allowed frames execute scripts (but other sandboxed properties still apply -- specifically, forms are disabled), and the disallowed ones do not. If successful the test prints "PASS".

                  
PASS
